Fixed bug related to ligands provided as MolFileEntry
In the LUNA pipeline, PDB molecules are converted to a molecular file, usually MOL. However, for ligands provided as MolFileEntry that is not necessary, and, for that reason, the conversion is skipped for those ligands.

However, there was a bug in the code that were forcing the ligand conversion from PDB to MOL when processing a residue. That happened because the ligand was being captured by the method 'get_proximal_compounds(comp)' at line #896, and then converted from PDB to MOL even if it was not necessary. 

Therefore, we modified the code so that the ligand is skipped at line #896 as well.