Faster version of perceive_atom_groups()
In the previous version, perceive_atom_groups() would loop through each compound in the list to convert it to MOL with its neighboring residues. This process was inefficient and slow.

To speed up the function, I modified the code to perform only one conversion per run, i.e., all residues in the binding site (ligand entries) or the entire structure (chain entries) are now converted to MOL at once. In a test I made, the average time to process an entry dropped from 4s to 1s with Open Babel.

However, I found out that the RDKit function to perceive pharmacophoric properties is extremely slow when working with large molecular files. For instance, in the same tests I made, the average time to process an entry increased from 7s to 24s. So, this new solution is only appropriate for Open Babel. 

Considering that Open Babel produces the same results as RDKit, I decided to remove the support for RDKit in this function. After all, RDKit was never used for converting PDBs to MOL. Under the hoods, LUNA always used Open Babel to convert PDBs to MOL. If the user sets 'mol_obj_type' to RDKit, LUNA would then convert the Open Babel object to RDKit and return it. Therefore, in the end, it didn't matter which library was chosen. This will only matter if an external molecular file is provided. In this case, RDKit is still an option, and pharmacophoric properties could be perceived with it.

To reflect all these updates, I removed the support for RDKit from biopython_entity_to_mol() by removing the argument 'mol_obj_type'. Additionally, The flags  'validate_mol' and 'standardize_mol' are now controlled by 'amend_mol'. Previously, 'standardize_mol' would be set to True only if the target compound was a residue. As perceive_atom_groups() now calls biopython_entity_to_mol() by providing a list of ligands and residues, it doesn't make sense to have two flags anymore.