Validator now fixes N atoms that forms a dative bond with metals
Added the function _fix_atoms_bound_to_metals(). At the current version, this function verifies if the N from a N-terminal residue, ND1/NE2 from HIS, or NZ from LYS has a dative bond with metals. If so, it checks if the atom is charged. If so, the Validator deprotonate it so that the dative bond is taken into consideration. That's necessary because Open Babel ignores the presence of the metal and protonates the nitrogens.