Fixed a bug related to the cache system
Fixed a bug in function cache_protein_properties():

For some complexes, different results were produced when using the cache system in comparison to when it was turned off. This was especially true for complexes having metal ions forming metal complexes. This happened in part because ligands and metals were being excluded from the cache processing.

Now, it permits such molecules to be part of the cache creation so that valences/bonds and metal complexes are correctly perceived and standardized, but all atom groups formed by those molecules are removed at the end because only residues and water should be saved in the cache.

By doing so, the bug was removed and now the complexes producing different results are now fixed.

OBS: there is only one exception: 1I2C:A:UPG:402. In this complex, there is one hydrogen bond that only appears when the cache is not used. The cause of such a bug is Open Babel, which places hydrogens for the atom TYR182/OH in different positions depending on the number of residues that compose the binding site during the conversion of PDB to MOL.